requirements.txt
articles==0.1.0
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.5
comtypes==1.1.11
contourpy==1.0.6
cycler==0.11.0
Django==4.1.1
django-admin-interface==0.24.2
django-bower==5.2.0
django-colorfield==0.8.0
django-scheduler==0.10.0
django-smart-selects==1.6.0
django-urls==1.1.3
fonttools==4.38.0
gunicorn==20.1.0
gym==0.26.2
gym-notices==0.0.8
icalendar==5.0.3
idna==3.3
joblib==1.1.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.6.2
numpy==1.23.0
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.3
pycparser==2.21
pyglet==1.5.21
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.1
pyttsx3==2.90
pytz==2022.6
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sounddevice==0.4.4
South==1.0.2
sqlparse==0.4.2
srt==3.5.2
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.64.0
tzdata==2022.2
urllib3==1.26.9
Voice==0.1.0
vosk==0.3.42
websockets==10.4
